Move android lib & gradle files into main tree
Due to being created via Android Studio, the android library was set up
as a standalone project. Restructuring this such that the android library is a
subdirectory of crosslib-core and the root gradle files are in the repository
root facilitates multiple projects (e.g. example android applications) and to
better manage the build process across the different libraries.

Signed-off-by: Ben Steadman <steadmanben1@gmail.com>