Basic example Android Application
To demonstrate usage of the Android library as well as building multiple
projects with Gradle, this commit introduces a simple Android
application *separate* to the crosslib-core Android library. The app
depends on the library (as if it were a third party) and is deliberately
simple; the point here is *not* to demonstrate a fully featured Android
application.

Signed-off-by: Ben Steadman <steadmanben1@gmail.com>